+++ scripts/template_shapes/WalledTemplateRectangle.js
@@ -45,6 +45,23 @@ export class WalledTemplateRectangle extends WalledTemplateCircle {
   }
 
   /**
+   * Compute the shape to be used for this template.
+   * Output depends on the specific template settings.
+   * @returns {PIXI.Polygon|PIXI.Circle}
+   */
+  computeShape() {
+    const shape = super.computeShape();
+
+    // Set values that Sequencer or other modules may use from the rectangle.
+    if ( !shape.width || !shape.height ) {
+      const origShape = this.originalShape;
+      shape.width ??= origShape.width;
+      shape.height ??= origShape.height;
+    }
+    return shape;
+  }
+
+  /**
    * Generate a new RectangleTemplate based on spreading from a designated corner.
    * @param {PIXI.Point} corner
    * @returns {WalledTemplateRectangle[]|null}